Fix Linux compilation

+++ Engine/RotoStrokeItem.cpp
@@ -596,7 +596,7 @@ RotoStrokeItem::getMostRecentStrokeChangesSinceAge(double time,
     
     if (lastAge == _imp->strokes[lastMultiStrokeIndex].xCurve->getKeyFramesCount() - 1) {
         //We rendered completly that stroke so far, pick the next one if there is
-        if (lastMultiStrokeIndex == _imp->strokes.size() - 1) {
+        if (lastMultiStrokeIndex == (int)_imp->strokes.size() - 1) {
             //nothing to do
             return false;
         } else {
